+++ core/apis_client.go
@@ -17,22 +17,28 @@ package core
 import (
 	"application/core/model"
 	"application/driven/uiucadapters"
+
 	"encoding/json"
 	"os"
 )
 
 // appClient contains client implementations
 type appClient struct {
-	app             *Application
-	Courseadapter   Courses
-	LocationAdapter WayFinding
-	LaundryAdapter  LaundryService
-	ContactAdapter  Contact
+	app                *Application
+	Courseadapter      Courses
+	LocationAdapter    WayFinding
+	LaundryAdapter     LaundryService
+	ContactAdapter     Contact
+	SuccessTeamAdapter SuccessTeam
 }
 
 // GetExample gets an Example by ID
 func (a appClient) GetExample(orgID string, appID string, id string) (*model.Example, error) {
 	return a.app.shared.getExample(orgID, appID, id)
+}
+
+func (a appClient) GetUnitCalendars(id string) (*[]model.UnitCalendar, error) {
+	return a.app.storage.FindCalendars(id)
 }
 
 func (a appClient) ListLaundryRooms() (*model.Organization, error) {
@@ -139,6 +145,47 @@ func (a appClient) GetTermSessions() (*[4]model.TermSession, error) {
 	return retData, nil
 }
 
+func (a appClient) GetSuccessTeam(uin string, unitid string, accesstoken string) (*model.SuccessTeam, int, error) {
+	conf, _ := a.app.GetEnvConfigs()
+
+	calendars, err := a.app.storage.FindCalendars(unitid)
+	if err != nil {
+		return nil, 500, err
+	}
+
+	retData, status, err := a.SuccessTeamAdapter.GetSuccessTeam(uin, calendars, accesstoken, conf)
+	if err != nil {
+		return nil, status, err
+	}
+	return retData, status, nil
+
+}
+
+func (a appClient) GetPrimaryCareProvider(uin string, accesstoken string) (*[]model.SuccessTeamMember, int, error) {
+	conf, _ := a.app.GetEnvConfigs()
+	retData, status, err := a.SuccessTeamAdapter.GetPrimaryCareProvider(uin, accesstoken, conf)
+	if err != nil {
+		return nil, status, err
+	}
+	return retData, status, nil
+}
+
+func (a appClient) GetAcademicAdvisors(uin string, unitid string, accesstoken string) (*[]model.SuccessTeamMember, int, error) {
+	conf, _ := a.app.GetEnvConfigs()
+
+	calendars, err := a.app.storage.FindCalendars(unitid)
+	if err != nil {
+		return nil, 500, err
+	}
+
+	retData, status, err := a.SuccessTeamAdapter.GetAcademicAdvisors(uin, calendars, accesstoken, conf)
+	if err != nil {
+		return nil, status, err
+	}
+	return retData, status, nil
+
+}
+
 // newAppClient creates new appClient
 func newAppClient(app *Application) appClient {
 
@@ -158,5 +205,6 @@ func newAppClient(app *Application) appClient {
 	client.LaundryAdapter = uiucadapters.NewCSCLaundryAdapter(laundryAssets)
 	client.Courseadapter = uiucadapters.NewCourseAdapter()
 	client.LocationAdapter = uiucadapters.NewUIUCWayFinding()
+	client.SuccessTeamAdapter = uiucadapters.NewSuccessTeamAdapter()
 	return client
 }