+++ core/interfaces.go
@@ -30,6 +30,7 @@ type Default interface {
 // Client exposes client APIs for the driver adapters
 type Client interface {
 	GetExample(orgID string, appID string, id string) (*model.Example, error)
+	GetUnitCalendars(id string) (*[]model.UnitCalendar, error)
 	ListLaundryRooms() (*model.Organization, error)
 	GetLaundryRoom(roomid string) (*model.RoomDetail, error)
 	InitServiceRequest(machineid string) (*model.MachineRequestDetail, error)
@@ -41,6 +42,9 @@ type Client interface {
 	GetGiesCourses(uin string, accessToken string) (*[]model.GiesCourse, int, error)
 	GetStudentCourses(uin string, termid string, accessToken string) (*[]model.Course, int, error)
 	GetTermSessions() (*[4]model.TermSession, error)
+	GetSuccessTeam(uin string, unitid string, accessToken string) (*model.SuccessTeam, int, error)
+	GetPrimaryCareProvider(uin string, accessToken string) (*[]model.SuccessTeamMember, int, error)
+	GetAcademicAdvisors(uin string, unitid string, accessToken string) (*[]model.SuccessTeamMember, int, error)
 }
 
 // Admin exposes administrative APIs for the driver adapters
@@ -106,6 +110,8 @@ type Storage interface {
 	UpdateExample(example model.Example) error
 	DeleteExample(orgID string, appID string, id string) error
 
+	FindCalendars(id string) (*[]model.UnitCalendar, error)
+
 	InsertLegacyEvents(context storage.TransactionContext, items []model.LegacyEventItem) error
 }
 
@@ -150,4 +156,11 @@ type Appointments interface {
 	CreateAppointment(appt *model.AppointmentPost, accesstoken string, conf *model.EnvConfigData) (*model.BuildingBlockAppointment, error)
 	DeleteAppointment(uin string, sourceid string, accesstoken string, conf *model.EnvConfigData) (string, error)
 	UpdateAppointment(appt *model.AppointmentPost, accesstoken string, conf *model.EnvConfigData) (*model.BuildingBlockAppointment, error)
+}
+
+// SuccessTeam represents the adapter needed to interface with the various assignedstaff end points to create the user's success team
+type SuccessTeam interface {
+	GetSuccessTeam(uin string, calendars *[]model.UnitCalendar, accesstoken string, conf *model.EnvConfigData) (*model.SuccessTeam, int, error)
+	GetPrimaryCareProvider(uin string, accesstoken string, conf *model.EnvConfigData) (*[]model.SuccessTeamMember, int, error)
+	GetAcademicAdvisors(uin string, calendars *[]model.UnitCalendar, accesstoken string, conf *model.EnvConfigData) (*[]model.SuccessTeamMember, int, error)
 }